Implement `CandleManager`
